Make Type Check In `validateAndExecute()` Stricter
We were seeing an issue in yarn v3.0.2 when using the workspaces plugin
where it was possible for a command's `schema` prop to be not
`undefined`, but also not an array. The issue is captured in the yarn
repository here: https://github.com/yarnpkg/berry/issues/3108

In order to ensure that `applyCascade()` is only called with valid
arguments, rather than checking that `cascade` is not `undefined`,
ensure that it is an `Array`. Otherwise, when we try to iterate over it,
we get a Type Error stating that `followups.every()` is not a function.

If `cascade` is defined, but isn't an array, then throw an error
notifying that the Command was constructed improperly with an invalid
`schema` property.